Add a little utility for timing fork()s.
There's an option to use rfork(), but RFNOWAIT doesn't appear to be
working as expected - lots of zombie processes are left over and don't
appear to get reaped.

The chain mode doesn't work at the moment since we don't have a way to
wait for the chain to finish. Probably we can designate a reaper that
waits for the requisite number of zombies.